requirements.txt
alembic==1.5.7
appdirs==1.4.4
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coverage==5.5
distlib==0.3.1
docopt==0.6.2
fabulous==0.3.0
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
Pillow==8.0.1
pipreqs==0.4.10
psycopg2-binary==2.8.6
pycparser==2.20
pyfiglet==0.8.post1
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.3
virtualenv==20.2.1
Werkzeug==1.0.1
yarg==0.1.9
